simplify compiler stuff
