+++ src/webpack/compiler/createCompiler.js
@@ -1,20 +1,9 @@
 // @flow
 import webpack from 'webpack';
-import type { Compiler, Stats } from '../types';
+import type { Compiler } from '../types';
 
-export default function createCompiler(webpackConfig: {}, cb: (err: ?{}, stats: Stats) => void): Compiler {
+export default function createCompiler(webpackConfig: {}): Compiler {
   const compiler = webpack(webpackConfig);
-
-  compiler.plugin('failed', cb);
-  compiler.plugin('done', (stats) => {
-    if (stats.hasErrors()) {
-      const jsonStats = stats.toJson();
-      const [err] = jsonStats.errors;
-      cb(err, null);
-    } else {
-      cb(null, stats);
-    }
-  });
 
   return compiler;
 }
